+++ custom_components/easywave/config_flow_device.py
@@ -7,7 +7,6 @@ from easywave_home_control.codec import (
     SensorLearnPayload,
     SensorTelegramEvent,
 )
-import voluptuous as vol
 
 from homeassistant.config_entries import SubentryFlowResult
 
@@ -308,17 +307,13 @@ class EasywaveDeviceAddFlowMixin(EasywaveDeviceFlowMixin):
                 title=user_input["title"],
                 unique_id=unique_id,
                 data=data,
+                area_id=self._area_id_from_input(user_input),
             )
 
         return self.async_show_form(
             step_id="transmitter_confirm",
-            data_schema=vol.Schema(
-                {
-                    vol.Required(
-                        "title",
-                        default=self._next_default_name(ENTRY_TYPE_TRANSMITTER),
-                    ): str,
-                }
+            data_schema=self._confirm_name_area_schema(
+                title_default=self._next_default_name(ENTRY_TYPE_TRANSMITTER),
             ),
         )
 
@@ -375,17 +370,13 @@ class EasywaveDeviceAddFlowMixin(EasywaveDeviceFlowMixin):
                 title=user_input["title"],
                 unique_id=unique_id,
                 data=data,
+                area_id=self._area_id_from_input(user_input),
             )
 
         return self.async_show_form(
             step_id="sensor_confirm",
-            data_schema=vol.Schema(
-                {
-                    vol.Required(
-                        "title",
-                        default=self._next_default_name(ENTRY_TYPE_NEO_SENSOR),
-                    ): str,
-                }
+            data_schema=self._confirm_name_area_schema(
+                title_default=self._next_default_name(ENTRY_TYPE_NEO_SENSOR),
             ),
             description_placeholders={
                 "sensor_list": await self._async_format_neo_sensor_list(